move RestApiDeployer class to internal.deployment.listener package
