+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/StartupListener.java
@@ -35,8 +35,8 @@ import org.wso2.carbon.deployment.engine.Deployer;
 import org.wso2.carbon.kernel.startupresolver.RequiredCapabilityListener;
 import org.wso2.carbon.kernel.startupresolver.StartupServiceUtils;
 import org.wso2.carbon.uis.api.ServerConfiguration;
-import org.wso2.carbon.uis.internal.deployment.RestApiDeployer;
 import org.wso2.carbon.uis.internal.deployment.listener.AppTransportBinder;
+import org.wso2.carbon.uis.internal.deployment.listener.RestApiDeployer;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroservicesRegistrar;
 import org.wso2.carbon.uis.internal.io.deployment.ArtifactAppDeployer;
 import org.wso2.carbon.uis.spi.RestApiProvider;
