+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/deployment/listener/RestApiDeployer.java
@@ -16,12 +16,13 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal.deployment;
+package org.wso2.carbon.uis.internal.deployment.listener;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.ServerConfiguration;
+import org.wso2.carbon.uis.internal.deployment.AppDeploymentEventListener;
 import org.wso2.carbon.uis.internal.exception.AppDeploymentEventListenerException;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroservicesRegistrar;
